Full backup now saves getExternalFilesDir() content with the app data
... instead of only saving it with the enormous "shared storage" generic
data blob.  In parallel, we no longer store managed app-specific files
on external storage as part of the generic shared-storage blob.

At restore time we just go ahead and apply such files, though,
because they're a priori going to be part of an archive generated by
an older version of the platform, so that's how the data is expected
to be handled in those circumstances.

Bug 6718844

Change-Id: I4410514d368b11d74b3afe6b92d363d4115a3415