Load crypto plugins from additional shared libraries
Currently crypto plugins are expected to be in libdrmdecrypt.so.
When there are multiple plugins supporting different schemes,
this approach requires source code integration across
vendors which is unmanagable.  Also, for integration with
MediaDrm where the crypto keys are obtained from a drm server,
the MediaCrypto plugin needs to interoperate with the MediaDrm
plugin. This change allows {MediaCrypto, MediaDrm} pairs that
are logically related to be implemented in a common shared
library.

Multi-repository commit, goes with:
https://googleplex-android-review.googlesource.com/287838

Change-Id: I08577cda0cbcb22ebe3cb874f5fcafe411c36be3