Fix an action bar menu order of operations bug
As options menu invalidations in PhoneWindow are deferred, it was
possible for a call to open the overflow menu (which involves an extra
chance to prepare the menu) to happen before the pending menu
invalidation was handled. Process any pending menu invalidations
before attempting to show the overflow menu in response to one of
these events.

Bug 7209308

Change-Id: Id50f7f6ad5d3a631745207fa0bad8c356690e16c