Stop jank from window animations overlapping activity animations
Relaunching a running activity will sometimes cause animations in that
activity to run, causing jank due to too much happening (window animations
on that same window running at the same time).

An earlier fix pauses application rendering while window animations are running,
but only the first time the activity comes up. If the window is animated
after that (such as is the case with re-launching it while it is running
in the background), rendering, and therefore animations, happen as usual,
causing the jank.

The fix is to simply broaden the scope of when application rendering is
paused to include anytime the window is animating.

Issue #8472972 Make app animations play nice with window animations

Change-Id: I1dd8da039fcb4d1faf6c0811bae97ef2847deb84