Fully setup newly bound service before state management.
If the connected service is not entirely setup when calling the method for
handling a change in the current user state we get a potential NPE since
the management method may have discarded the service, thus nullifying the
connection to it. Now the service is fully configured before calling the
state change management method.

bug:8600489

Change-Id: Ib0bf7c6d575e15c620da419d43ece22f4187fd34