Magnification does not work if no enabled accessibility services.
This is a regression in which the input filter of the accessibility
manager service is not set if magnification is enabled but accessibility
is not - i.e. no accessibility serivces are enabled. Fixed the logic to
install the input filter if magnification is on but services are not
enabled in addition to services being enabled.

bug:8652765

Change-Id: Ia73e1064035f95ba0f246f4cabcc42d58c12a11f